Conflicts:
	CMakeLists.txt
	Sources/ArxCommon/ARX_Common.cpp
	Sources/Athena/Athena.cpp
	Sources/Athena/Athena_Ambiance.cpp
	Sources/Athena/Athena_CoDec_ADPCM.cpp
	Sources/Athena/Athena_Codec_ADPCM.h
	Sources/Athena/Athena_Codec_RAW.cpp
	Sources/Athena/Athena_Codec_RAW.h
	Sources/Athena/Athena_DLL.cpp
	Sources/Athena/Athena_Environment.h
	Sources/Athena/Athena_FileIO.cpp
	Sources/Athena/Athena_Global.cpp
	Sources/Athena/Athena_Global.h
	Sources/Athena/Athena_Instance.cpp
	Sources/Athena/Athena_Instance.h
	Sources/Athena/Athena_Mixer.cpp
	Sources/Athena/Athena_Resource.h
	Sources/Athena/Athena_Sample.cpp
	Sources/Athena/Athena_Stream_ASF.cpp
	Sources/Athena/Athena_Stream_ASF.h
	Sources/Athena/Athena_Stream_WAV.cpp
	Sources/Athena/Athena_Stream_WAV.h
	Sources/Athena/Athena_Track.cpp
	Sources/Athena/Athena_Track.h
	Sources/DANAE_Debugger/SCRIPT_DEBUGGER_Dialog.cpp
	Sources/Include/ARX_CCinematique.h
	Sources/Include/ARX_Draw.h
	Sources/Include/ARX_NPC.h
	Sources/Include/ARX_Network.h
	Sources/Include/ARX_Sound.h
	Sources/Include/ARX_Special.h
	Sources/Include/ARX_Speech.h
	Sources/Include/ARX_Spells.h
	Sources/Include/Athena.h
	Sources/Include/Athena_Types.h
	Sources/Include/EERIEDraw.h
	Sources/Include/EERIEMath.h
	Sources/Include/EERIEPoly.h
	Sources/Include/EERIEProgressive.h
	Sources/Include/EERIETypes.h
	Sources/Include/EERIE_AVI.h
	Sources/Include/HERMESMain.h
	Sources/Include/HERMESNet.h
	Sources/Include/HERMES_pack_public.h
	Sources/Include/Mercury_extern.h
	Sources/Include/OtherLibs/implode.h
	Sources/Include/OtherLibs/jmemsys.h
	Sources/Mercury/Mercury_global.cpp
	Sources/danae/ARX_Loc.cpp
	Sources/danae/ARX_Sound.cpp
	Sources/danae/ARX_Speech.cpp
	Sources/danae/ARX_SpellFx_Lvl01.h
	Sources/danae/ARX_ViewImage.cpp
	Sources/danae/DanaeDlg.cpp
	Sources/eerie/EERIEApp.cpp
	Sources/eerie/EERIEClothes.cpp
	Sources/eerie/EERIEDraw.cpp
	Sources/eerie/EERIELinkedObj.cpp
	Sources/eerie/EERIEPoly.cpp
	Sources/eerie/EERIE_AVI.cpp
	Sources/hermes/HERMESmain.cpp

+++ Sources/Include/Athena_Types.h
@@ -41,8 +41,8 @@ namespace ATHENA
 	typedef signed char aalSByte;
 	typedef unsigned short aalUWord;
 	typedef signed short aalSWord;
-	typedef unsigned long aalULong;
-	typedef signed long aalSLong;
+	typedef unsigned int aalULong;
+	typedef signed int aalSLong;
  
 	typedef signed aalSInt;
 	typedef float aalFloat;
